EmptyHash -> NullHash Implemented Replay/Verifier, not yet tested.
git-svn-id: https://svn.mpi-sws.org/DS/freepastry/trunk@3891 c21dacaf-8602-0410-aacd-8bc41df2cbe7

+++ pastry/src/org/mpisws/p2p/transport/peerreview/history/SecureHistory.java
@@ -120,7 +120,7 @@ public interface SecureHistory /* extends Iterable<IndexEntry> */ {
   /**
    *  Get the content of a log entry, specified by its record number 
    */
-  public byte[] getEntry(int idx, int maxSizeToRead) throws IOException;
+  public byte[] getEntry(long idx, int maxSizeToRead) throws IOException;
 
   /**
    * If the log already contains an entry in 'hashed' form and we learn the actual
